refactor: move permissions helpers to separate file
