+++ dev-client/src/util.ts
@@ -16,7 +16,6 @@
  */
 import {NativeModules, Platform} from 'react-native';
 
-import {SiteUserRole} from 'terraso-client-shared/selectors';
 import {Coords} from 'terraso-client-shared/types';
 import {
   isValidLatitude,
@@ -145,21 +144,6 @@ export const isValidCoordinates = (input: string) => {
 
   return isValidLatitude(latitude) && isValidLongitude(longitude);
 };
-
-export const matchesOne =
-  <T>(cmp: (a: T, b: T) => boolean) =>
-  (examples: T[]) =>
-  (sample: T) =>
-    examples.filter(x => cmp(x, sample)).length > 0;
-
-export const matchesRole = matchesOne(
-  (a: SiteUserRole, b: SiteUserRole) => a.kind === b.kind && a.role === b.role,
-);
-
-export const isSiteManager = matchesRole([
-  {kind: 'site', role: 'OWNER'},
-  {kind: 'project', role: 'MANAGER'},
-]);
 
 export const getSoilWebUrl = (coords: Coords) => {
   return `https://casoilresource.lawr.ucdavis.edu/gmap/?loc=${coords.latitude},${coords.longitude}`;
